backend/pyproject.toml
[project]
name = "file-manager-api"
version = "0.1.0"
description = "Simple file management API using FastAPI and GCP"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic-settings>=2.2",
  "firebase-admin>=6.4",
  "google-cloud-storage>=2.14",
  "google-cloud-firestore>=2.14",
  "python-multipart>=0.0.9",
  "prometheus-fastapi-instrumentator>=7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
]

[tool.uvicorn]
factory = false
